Pico can now read and interpret commands over UART. Functionality to display data on screen successful. Not tested on RT.

+++ pico/src/display.h
@@ -1,11 +1,10 @@
+#include <stdint.h>
 #include <stdio.h>
 #include "pico/stdlib.h"
 
-
 void init_display();
-void update_RPM(int *rpm);
-void update_gear(int *gearNumber);
-void update_speed(int *speed);
+void update_RPM(uint16_t * rpm);
+void update_gear(uint16_t * gearNumber);
+void update_speed(uint16_t * speed);
 void draw_display();
-int findRevMeterPixels(int *engineRPM);
-
+int findRevMeterPixels(uint16_t * engineRPM);
